+++ inscripta/biocantor/io/gff3/parser.py
@@ -53,9 +53,12 @@ class GffutilsParseArgs:
     merge_strategy: Optional[str] = "create_unique"
 
 
-def filter_and_sort_qualifiers(qualifiers: Dict[str, List[str]]) -> Dict[str, List[str]]:
-    """Filter out the qualifiers for any terms we have elevated"""
-    return {key: sorted(vals) for key, vals in qualifiers.items() if not re.match(BIOCANTOR_QUALIFIERS_REGEX, key)}
+def filter_and_sort_qualifiers(qualifiers: Dict[str, List[str]]) -> Optional[Dict[str, List[str]]]:
+    """Filter out the qualifiers for any terms we have elevated and also any GFF3 special terms"""
+    qualifiers = {
+        key: sorted(vals) for key, vals in qualifiers.items() if not re.match(BIOCANTOR_QUALIFIERS_REGEX, key)
+    }
+    return qualifiers if qualifiers else None
 
 
 def _parse_genes(chrom: str, db: FeatureDB) -> List[Dict]:
@@ -209,10 +212,14 @@ def _find_all_top_level_non_gene_features(chrom: str, db: FeatureDB, feature_typ
             yield feature
 
 
-def _parse_child_features_to_feature_interval(features: List[Feature]) -> Dict[str, Any]:
+def _parse_child_features_to_feature_interval(
+    features: List[Feature], locus_tag: Optional[str] = None
+) -> Dict[str, Any]:
     """
     Extract values from a list of child features and produce a dictionary to build a
     :class:`~biocantor.io.models.FeatureIntervalModel` from.
+
+    Can also be provided a ``locus_tag`` value from a parent, if applicable.
 
     This function combines all child features of a top-level non-gene feature
     """
@@ -225,7 +232,6 @@ def _parse_child_features_to_feature_interval(features: List[Feature]) -> Dict[s
     feature_ids = Counter()
     strand = None
     chrom = None
-    locus_tag = None
     interval_starts = []
     interval_ends = []
     for feature in features:
@@ -252,8 +258,6 @@ def _parse_child_features_to_feature_interval(features: List[Feature]) -> Dict[s
             this_locus_tag = feature.attributes[BioCantorQualifiers.LOCUS_TAG.value][0]
             if this_locus_tag != locus_tag:
                 raise GFF3LocusTagError("Cannot have multiple child features with different locus tags.")
-            elif not locus_tag:
-                locus_tag = this_locus_tag
 
         if not qualifiers:
             qualifiers = feature.attributes
@@ -278,7 +282,6 @@ def _parse_child_features_to_feature_interval(features: List[Feature]) -> Dict[s
         feature_id=feature_id,
         feature_name=feature_name,
         feature_types=sorted(feature_types),
-        locus_tag=locus_tag,
         sequence_name=chrom,
         is_primary_feature=False,
     )
@@ -302,6 +305,11 @@ def _parse_features(chrom: str, db: FeatureDB, feature_types: List[str]) -> List
     for top_level_feature in _find_all_top_level_non_gene_features(chrom, db, feature_types):
         children = list(db.children(top_level_feature, level=1))
 
+        # extract parent locus tag to compare to children
+        locus_tag = None
+        if BioCantorQualifiers.LOCUS_TAG.value in top_level_feature.attributes:
+            locus_tag = top_level_feature.attributes[BioCantorQualifiers.LOCUS_TAG.value][0]
+
         if not children:
             # treat this isolated feature as both FeatureIntervalCollection and FeatureInterval
             feature = _parse_child_features_to_feature_interval([top_level_feature])
@@ -310,33 +318,31 @@ def _parse_features(chrom: str, db: FeatureDB, feature_types: List[str]) -> List
                 feature_intervals=[feature],
                 feature_collection_name=feature["feature_name"],
                 feature_collection_id=feature["feature_id"],
-                locus_tag=feature["locus_tag"],
+                locus_tag=locus_tag,
                 sequence_name=chrom,
+                qualifiers=feature["qualifiers"],
             )
+            # remove qualifiers from feature
+            del feature["qualifiers"]
         else:
             # combine all children into a FeatureInterval
-            feature = _parse_child_features_to_feature_interval(children)
-            feature__collection_name, feature__collection_id = extract_feature_name_id(top_level_feature.attributes)
-
-            locus_tag = None
-            if BioCantorQualifiers.LOCUS_TAG.value in top_level_feature.attributes:
-                locus_tag = top_level_feature.attributes[BioCantorQualifiers.LOCUS_TAG.value][0]
-            if locus_tag != feature["locus_tag"]:
-                raise GFF3LocusTagError("Locus tag of parent feature did not match children.")
+            feature = _parse_child_features_to_feature_interval(children, locus_tag=locus_tag)
+            feature_collection_name, feature_collection_id = extract_feature_name_id(top_level_feature.attributes)
 
             feature_collection = dict(
                 feature_intervals=[feature],
-                feature_collection_name=feature__collection_name,
-                feature_collection_id=feature__collection_id,
+                feature_collection_name=feature_collection_name,
+                feature_collection_id=feature_collection_id,
                 locus_tag=locus_tag,
                 sequence_name=chrom,
+                qualifiers=filter_and_sort_qualifiers(top_level_feature.attributes),
             )
 
         feature_collections.append(feature_collection)
     return feature_collections
 
 
-def _find_non_gene_feature_types(db: FeatureDB) -> List[str]:
+def _find_non_gene_feature_types(db: FeatureDB, feature_types_to_ignore: Optional[Set[str]] = None) -> List[str]:
     """Non-gene feature types are those that are not either a member of :class:`~biocantor.gene.biotype.Biotype`
     or :class:`~biocantor.io.gff3.constants.GFF3GeneFeatureTypes`. This combination of filters prevents genes being
     inadvertently pulled in from either of the two main styles of representing them.
@@ -345,7 +351,9 @@ def _find_non_gene_feature_types(db: FeatureDB) -> List[str]:
     Ensembl/GENCODE Style: gene -> transcript -> {exon, cds}
 
     Args:
-        db: db: Database from :mod:`gffutils`.
+        db: Database from :mod:`gffutils`.
+        feature_types_to_ignore: Feature types to ignore, if chosen. This is often used to ignore pointless features
+            like chromosome representations.
 
     Returns:
         A list of strings representing the non-gene feature types found in the database.
@@ -353,6 +361,8 @@ def _find_non_gene_feature_types(db: FeatureDB) -> List[str]:
     non_gene_feature_types = []
     for feat_type in db.featuretypes():
         if GFF3GeneFeatureTypes.has_value(feat_type):
+            continue
+        elif feature_types_to_ignore and feat_type in feature_types_to_ignore:
             continue
         else:
             try: