+++ inscripta/biocantor/io/models.py
@@ -47,7 +47,6 @@ class FeatureIntervalModel(BaseModel):
     feature_types: Optional[List[str]] = None
     feature_name: Optional[str] = None
     feature_id: Optional[str] = None
-    locus_tag: Optional[str] = None
     is_primary_feature: Optional[bool] = None
 
     def to_feature_interval(
@@ -77,7 +76,6 @@ class FeatureIntervalModel(BaseModel):
             feature_types=self.feature_types,
             feature_name=self.feature_name,
             feature_id=self.feature_id,
-            locus_tag=self.locus_tag,
             guid=self.feature_interval_guid,
             feature_guid=self.feature_guid,
             is_primary_feature=self.is_primary_feature,